+++ src/Entity/PortEntity.cs
@@ -4,7 +4,7 @@ public class PortEntity
 {
    public required string PortId { get; set; }
     public required string ConnectorType { get; set; }
-    public required float PowerKw { get; set; }
+    public required double PowerKw { get; set; }
     public required string ChargingMechanism { get; set; }
     public required string Status { get; set; }
     public required DateTime LastUpdated { get; set; }
